packages/stateful-components/src/inputs/editor.tsx
import React from "react";
import { connect } from "react-redux";
import { Dispatch } from "redux";

import { actions, AppState, ContentRef, selectors } from "@nteract/core";

export interface PassedEditorProps {
  id: string;
  contentRef: ContentRef;
  editorType: string;
  editorFocused: boolean;
  value: string;
  channels: any;
  kernelStatus: string;
  theme: string;
  onChange: (text: string) => void;
  onFocusChange: (focused: boolean) => void;
  className: string;
}

export interface EditorSlots {
  [key: string]: (props: PassedEditorProps) => React.ReactNode;
}

interface ComponentProps {
  id: string;
  contentRef: ContentRef;
  children: EditorSlots;
}

interface StateProps {
  editorType: string;
  editorFocused: boolean;
  value: string;
  channels: any;
  kernelStatus: string;
  theme: string;
}

interface DispatchProps {
  onChange: (text: string) => void;
  onFocusChange: (focused: boolean) => void;
}

type Props = ComponentProps & StateProps & DispatchProps;

export class Editor extends React.PureComponent<Props> {
  render(): React.ReactNode {
    const { editorType, children } = this.props;

    const chosenEditor = children ? children[editorType] : undefined;

    if (chosenEditor) {
      return chosenEditor({
        id: this.props.id,
        contentRef: this.props.contentRef,
        editorType: this.props.editorType,
        value: this.props.value,
        editorFocused: this.props.editorFocused,
        channels: this.props.channels,
        kernelStatus: this.props.kernelStatus,
        theme: this.props.theme,
        onChange: this.props.onChange,
        onFocusChange: this.props.onFocusChange,
        className: "nteract-cell-editor"
      });
    }
    return null;
  }
}

export const makeMapStateToProps = (
  initialState: AppState,
  ownProps: ComponentProps
) => {
  const { id, contentRef } = ownProps;
  const mapStateToProps = (state: AppState): StateProps => {
    const model = selectors.model(state, { contentRef });

    let editorFocused = false;
    let channels = null;
    let kernelStatus = "not connected";
    let value = "";
    const editorType = selectors.editorType(state);
    const theme = selectors.userTheme(state);

    if (model && model.type === "notebook") {
      const cell = selectors.notebook.cellById(model, { id });
      if (cell) {
        editorFocused = model.editorFocused === id;
        value = cell.get("source", "");
        if (cell.cell_type === "code") {
          const kernel = selectors.kernelByContentRef(state, { contentRef });
          if (kernel) {
            channels = kernel.channels;
            kernelStatus = kernel.status || "not connected";
          }
        }
      }
    }

    return {
      editorFocused,
      value,
      channels,
      kernelStatus,
      editorType,
      theme
    };
  };

  return mapStateToProps;
};

export const makeMapDispatchToProps = (
  initialDispatch: Dispatch,
  ownProps: ComponentProps
) => {
  const { id, contentRef } = ownProps;
  const mapDispatchToProps = (dispatch: Dispatch) => {
    return {
      onChange: (text: string) => {
        dispatch(actions.updateCellSource({ id, value: text, contentRef }));
      },

      onFocusChange(focused: boolean): void {
        if (focused) {
          dispatch(actions.focusCellEditor({ id, contentRef }));
          // Assume we can focus the cell if now focusing the editor
          // If this doesn't work, we need to go back to checking !cellFocused
          dispatch(actions.focusCell({ id, contentRef }));
        }
      }
    };
  };
  return mapDispatchToProps;
};

export default connect(makeMapStateToProps, makeMapDispatchToProps)(Editor);
